Support fractional time zone offsets with decimal index
Changed DisplayTimeZoneIndex from int to decimal across config, models, and view models to support fractional UTC offsets. Updated time zone handling in TimeZoneProvider and DasBlogSettings to use NodaTime for accurate conversions. Refactored tests and mocks for new logic. Reformatted config files and updated related classes for consistency.

+++ source/DasBlog.Services/ConfigFile/Interfaces/ISiteConfig.cs
@@ -108,7 +108,7 @@ namespace DasBlog.Services.ConfigFile.Interfaces
 
 		bool ShowItemSummaryInAggregatedViews { get; set; }
 
-        int DisplayTimeZoneIndex { get; set; }
+        decimal DisplayTimeZoneIndex { get; set; }
 
         bool AdjustDisplayTimeZone { get; set; }
